Windows warning fix

+++ src/prodbg/ui/bgfx/bgfx_plugin_ui.cpp
@@ -739,7 +739,7 @@ PluginUI::State BgfxPluginUI::updateInstance(ViewPluginInstance* instance, PDRea
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-float BgfxPluginUI::getStatusBarSize()
+int BgfxPluginUI::getStatusBarSize()
 {
     return m_statusSize;
 }
